Reorg - move FetchURL to util; get rid of cli and syntax Util classes
